refactor: extract the generated column resolution into its own action
Which fields a component is generated with, and their types, was split
between the two GetStubVars actions. Moving it into ResolveGeneratedColumns
gives both a single source of truth to read from, and makes the resolution
reusable outside of stub rendering.
